refactor: shrunk BookCard's code lines for refactoring purposes
