+++ src/components/book-card/BookCard.tsx
@@ -1,13 +1,8 @@
 import {
-  useMutation,
   type RefetchOptions,
   type QueryObserverResult
 } from '@tanstack/react-query'
-import {
-  addBookToShelf,
-  removeFromShelf,
-} from '@services/bookshelves'
-import { AxiosError, type AxiosResponse } from 'axios'
+import type { AxiosResponse } from 'axios'
 import type { BookCardProps } from '@@types/bookCard'
 import type { MouseEventHandler } from 'react'
 import { AppLink } from "@components/ui/AppLink"
@@ -16,25 +11,8 @@ import { Button } from '@components/ui/button'
 import { ChevronRight } from 'lucide-react'
 import { Image } from "@components/ui/image"
 import { Title } from "@components/ui/title"
+import { useBookshelfHandler } from '@hooks/useBookshelfHandler'
 import { useLocation } from 'react-router-dom'
-import { toast } from 'sonner'
-
-type ToastProps = {
-  type?: 'success' | 'error'
-  title: string
-  description?: string
-}
-
-const triggerToast = ({ title, type = "success", description }: ToastProps) => {
-  toast[type](title, {
-    description,
-    duration: 3000,
-    action: {
-      label: 'Dismiss',
-      onClick: () => toast.dismiss(),
-    }
-  })
-}
 
 export const BookCard = ({
   title,
@@ -51,52 +29,24 @@ export const BookCard = ({
 ) => {
   const { pathname } = useLocation()
   const isYourBooks = pathname.includes('/your-books')
-
-  const { mutateAsync: addToShelfAsync } = useMutation({
-    mutationKey: ['booksIds', slug],
-    mutationFn: () => addBookToShelf(slug),
-  })
-
-  const { mutateAsync: removeFromShelfAsync } = useMutation({
-    mutationKey: ['removeFromShelf', slug],
-    mutationFn: (slug: string) => removeFromShelf(slug),
+  const shelfMutation = useBookshelfHandler({
+    isAdded,
+    slug,
+    refetch
   })
 
   const handleClick: MouseEventHandler<HTMLButtonElement> = async (e) => {
     e.preventDefault()
     e.stopPropagation()
-    try {
-      if (isAdded) {
-        await removeFromShelfAsync(slug)
-        triggerToast({
-          title: 'Book removed from shelf',
-          type: 'success',
-          description: 'You can always add it back later!'
-        })
-      } else {
-        await addToShelfAsync()
-        triggerToast({
-          title: 'Book added to shelf',
-          type: 'success',
-          description: 'Go to your shelf to check it out!'
-        })
-      }
-    } catch (error) {
-      if (error instanceof AxiosError) {
-        console.error('Error updating book shelf:', error.message)  
-      }
-      triggerToast({
-        title: 'Error updating book shelf',
-        type: 'error',
-        description: 'Please try again later.'
-      })
-    }
-    
-    if (refetch) await refetch()
+    await shelfMutation()
   }
 
   return (
-    <article className="w-full flex flex-col items-start gap-3 group">
+    <article 
+      role="region"
+      aria-labelledby={`book-title-${slug}`}
+      className="w-full flex flex-col items-start gap-3 group"
+    >
       <AppLink
         to={isYourBooks ? `/your-books/books/${slug}` : `/discover/books/${slug}`}
         className="w-full"
@@ -113,6 +63,7 @@ export const BookCard = ({
             <p className="text-xs text-primary font-semibold">{genre}</p>
             <AppLink to={`/books/${slug}`} className="w-full">
               <Title
+                id={`book-title-${slug}`}
                 as="h3"
                 className="text-base line-clamp-1 font-medium text-primary transition-colors hover:underline"
               >
@@ -124,7 +75,8 @@ export const BookCard = ({
           <Button
             variant="none"
             size="none"
-            className="hover:cursor-pointer lg:hidden"
+            className="sr-only hover:cursor-pointer lg:hidden"
+            aria-label={isAdded ? 'Remove from bookshelf' : 'Add to bookshelf'}
             onClick={handleClick}
           >
             {isAdded ? 'Remove' : 'Add'}
@@ -137,7 +89,7 @@ export const BookCard = ({
             className="mt-3 w-fit group inline-flex items-center gap-0.5 text-sm font-medium text-primary transition-all hover:underline"
           >
             Keep reading
-            <ChevronRight className="text-primary size-4 duration-100 group-hover:translate-x-1" />
+            <ChevronRight aria-hidden="true" className="text-primary size-4 duration-100 group-hover:translate-x-1" />
           </AppLink>
         )}
       </section>
